requirements.txt
certifi==2017.11.5
chardet==3.0.4
idna==2.6
multidict==3.3.2
PyYAML==3.12
requests==2.18.4
six==1.11.0
urllib3==1.22
vcrpy==1.11.1
wrapt==1.10.11
yarl==0.16.0
